Add authToken usage when using script ./get-results-certifications.js + BSR moved test in a proprer file
